comments

+++ mlx/backend/common/utils.h
@@ -48,12 +48,12 @@ std::tuple<std::vector<int>, std::vector<std::vector<int64_t>>>
 collapse_contiguous_dims(
     const std::vector<int>& shape,
     const std::vector<std::vector<int64_t>>& strides,
-    int64_t size_cap = std::numeric_limits<int64_t>::max());
+    int64_t size_cap = std::numeric_limits<int32_t>::max());
 std::tuple<std::vector<int>, std::vector<std::vector<size_t>>>
 collapse_contiguous_dims(
     const std::vector<int>& shape,
     const std::vector<std::vector<size_t>>& strides,
-    size_t size_cap = std::numeric_limits<size_t>::max());
+    size_t size_cap = std::numeric_limits<int32_t>::max());
 
 inline std::tuple<std::vector<int>, std::vector<std::vector<size_t>>>
 collapse_contiguous_dims(
